no_std for guillotiere and guillotiere_ffi
