+++ .gitignore
@@ -1,4 +1,5 @@
-
 target
 corpus
 artifacts
+
+.idea
